% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{dice}
\alias{dice}
\title{a function to compute probability}
\usage{
dice(n, k)
}
\arguments{
\item{n}{there are n dices}

\item{k}{all number of the dices are k}
}
\value{
a probability
}
\description{
when there are n dices and all number of the dices are k,what is the probability?
}
\examples{
\dontrun{
dice(5,18)}
}
