% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{sq}
\alias{sq}
\title{sensitive question}
\usage{
sq(N, n, p, m1, m2)
}
\arguments{
\item{N}{investigate N persons}

\item{n}{there are n persons to say yes among N persons}

\item{p}{the probability of saying yes about a easy and public question}

\item{m1}{the number of white balls}

\item{m2}{the number of black balls}
}
\value{
a probability
}
\description{
a function to compute the pribability of sensitive question
}
\examples{
\dontrun{
sq(1583,389,0.5,20,30)
}
}
