\name{StatComp20045-package}
\alias{StatComp20045-package}
\alias{StatComp20045}
\docType{package}
\title{
   for 'Statistical Computing' course
}
\description{
  Several R and Rcpp function are included in this package for illustration.
}
\author{
Chen Hao Cui

Maintainer: Chen Hao Cui <cuichenhao10@163.com>
}

\keyword{ package }

